Split tooltip view and type
